requirements.txt
streamlit
google-generativeai
edge-tts
pydub
python-dotenv
pillow
pytest
pytest-asyncio
gTTS
requests
audioop-lts
